✨ Add navigation support for the future
