+++ App.js
@@ -1,22 +1,18 @@
 import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View } from "react-native";
+import { StyleSheet, View } from "react-native";
 import LoginForm from "./login-screen/login-form";
+import { NavigationContainer } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+
+const stack = createNativeStackNavigator();
 
 export default function App() {
-  return (
-    <View style={styles.container}>
-      <StatusBar />
-      <LoginForm />
-    </View>
-  );
+    return (
+        <NavigationContainer>
+          <StatusBar />
+          <stack.Navigator>
+            <stack.Screen name="Login" component={LoginForm} options={{headerShown: false}} />
+            </stack.Navigator>
+        </NavigationContainer>
+    );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    borderColor: "#ff0000",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
