+++ package.json
@@ -10,8 +10,10 @@
   },
   "dependencies": {
     "@expo/webpack-config": "^19.0.0",
+    "@react-navigation/native-stack": "^6.9.13",
     "@types/react": "~18.2.14",
     "expo": "~49.0.8",
+    "expo-image": "~1.3.3",
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
@@ -19,8 +21,7 @@
     "react-native-paper": "^5.10.4",
     "react-native-web": "~0.19.6",
     "react-navigation": "^4.4.4",
-    "typescript": "^5.1.3",
-    "expo-image": "~1.3.3"
+    "typescript": "^5.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
